+++ src/cQueue.h
@@ -1,28 +1,44 @@
 /*!\file cQueue.h
 ** \author SMFSW
-** \copyright BSD 3-Clause License (c) 2017, SMFSW
+** \copyright BSD 3-Clause License (c) 2017-2018, SMFSW
 ** \brief Queue handling library (designed in c on STM32)
 ** \details Queue handling library (designed in c on STM32)
 **/
-
-#ifndef __CQUEUE_H__
-#define __CQUEUE_H__
-
+/****************************************************************/
+#ifndef __CQUEUE_H
+	#define __CQUEUE_H
 
 #ifdef __cplusplus
 extern "C"{
 #endif
+/****************************************************************/
+
 
 #include <inttypes.h>
 #include <stdbool.h>
 
 
+#define QUEUE_INITIALIZED	0x5AA5							//!< Queue initialized control value
+
+#define q_init_def(q, sz)	q_init(q, sz, 20, FIFO, false)	//!< Some kind of average default for queue initialization
+
+#define q_pull				q_pop							//!< \deprecated q_pull was already used in cQueue lib, alias is made to keep compatibility with earlier versions
+#define q_nbRecs			q_getCount						//!< \deprecated q_nbRecs was already used in cQueue lib, alias is made to keep compatibility with earlier versions
+#define q_clean				q_flush							//!< \deprecated q_clean was already used in cQueue lib, alias is made to keep compatibility with earlier versions
+
+
+/*!\enum enumQueueType
+** \brief Queue behavior enumeration (FIFO, LIFO)
+**/
 typedef enum enumQueueType {
-	FIFO = 0,
-	LIFO = 1
+	FIFO = 0,	//!< First In First Out behavior
+	LIFO = 1	//!< Last In First Out behavior
 } QueueType;
 
 
+/*!\struct Queue_t
+** \brief Queue type structure holding all variables to handle the queue
+**/
 typedef struct Queue_t {
 	QueueType	impl;		//!< Queue implementation: FIFO LIFO
 	bool		ovw;		//!< Overwrite previous records when queue is full allowed
@@ -33,14 +49,9 @@ typedef struct Queue_t {
 	uint16_t	in;			//!< number of records pushed into the queue
 	uint16_t	out;		//!< number of records pulled from the queue (only for FIFO)
 	uint16_t	cnt;		//!< number of records not retrieved from the queue
-	uint16_t	init;		//!< sets to 0x5A5A after a first init of the queue
+	uint16_t	init;		//!< set to QUEUE_INITIALIZED after successful init of the queue and reset when killing queue
 } Queue_t;
 
-
-#define q_init_def(q, sz)	q_init(q, sz, 20, FIFO, false)	//!< Some kind of average default for queue initialization
-
-#define q_pull				q_pop							//!< As pull was already used in SMFSW libs, alias is made to keep compatibility with earlier versions
-#define q_flush				q_clean							//!< As flush is a common keyword, alias is made to empty queue
 
 /*!	\brief Queue initialization
 **	\param [in,out] q - pointer of queue to handle
@@ -57,10 +68,19 @@ void * q_init(Queue_t * q, const uint16_t size_rec, const uint16_t nb_recs, cons
 **/
 void q_kill(Queue_t * q);
 
-/*!	\brief Clean queue, restarting from empty queue
+/*!	\brief Flush queue, restarting from empty queue
 **	\param [in,out] q - pointer of queue to handle
 **/
-void q_clean(Queue_t * q);
+void q_flush(Queue_t * q);
+
+/*!	\brief get initialization state of the queue
+**	\param [in] q - pointer of queue to handle
+**	\return Queue initialization status
+**	\retval true if queue is allocated
+**	\retval false is queue is not allocated
+**/
+inline bool __attribute__((always_inline)) q_isInitialized(const Queue_t * q) {
+	return (q->init == QUEUE_INITIALIZED) ? true : false; }
 
 /*!	\brief get emptiness state of the queue
 **	\param [in] q - pointer of queue to handle
@@ -69,8 +89,7 @@ void q_clean(Queue_t * q);
 **	\retval false is not empty
 **/
 inline bool __attribute__((always_inline)) q_isEmpty(const Queue_t * q) {
-	return (!q->cnt) ? true : false;
-}
+	return (!q->cnt) ? true : false; }
 
 /*!	\brief get fullness state of the queue
 **	\param [in] q - pointer of queue to handle
@@ -79,22 +98,20 @@ inline bool __attribute__((always_inline)) q_isEmpty(const Queue_t * q) {
 **	\retval false is not full
 **/
 inline bool __attribute__((always_inline)) q_isFull(const Queue_t * q) {
-	return (q->cnt == q->rec_nb) ? true : false;
-}
+	return (q->cnt == q->rec_nb) ? true : false; }
 
 /*!	\brief get number of records in the queue
 **	\param [in] q - pointer of queue to handle
 **	\return Number of records left in the queue
 **/
-inline uint16_t __attribute__((always_inline)) q_nbRecs(const Queue_t * q) {
-	return q->cnt;
-}
+inline uint16_t __attribute__((always_inline)) q_getCount(const Queue_t * q) {
+	return q->cnt; }
 
 /*!	\brief Push record to queue
 **	\param [in,out] q - pointer of queue to handle
 **	\param [in] record - pointer to record to be pushed into queue
 **	\return Push status
-**	\retval true if succefully pushed into queue
+**	\retval true if successfully pushed into queue
 **	\retval false if queue is full
 **/
 bool q_push(Queue_t * q, const void * record);
@@ -103,7 +120,7 @@ bool q_push(Queue_t * q, const void * record);
 **	\param [in] q - pointer of queue to handle
 **	\param [in,out] record - pointer to record to be popped from queue
 **	\return Pop status
-**	\retval true if succefully pulled from queue
+**	\retval true if successfully pulled from queue
 **	\retval false if queue is empty
 **/
 bool q_pop(Queue_t * q, void * record);
@@ -112,22 +129,24 @@ bool q_pop(Queue_t * q, void * record);
 **	\param [in] q - pointer of queue to handle
 **	\param [in,out] record - pointer to record to be peeked from queue
 **	\return Peek status
-**	\retval true if succefully pulled from queue
+**	\retval true if successfully pulled from queue
 **	\retval false if queue is empty
 **/
 bool q_peek(Queue_t * q, void * record);
 
-	/*!	\brief Drop current record from queue
-	**	\param [in,out] q - pointer of queue to handle
-	**	\return drop status
-	**	\retval true if succefully dropped from queue
-	**	\retval false if queue is empty
-	**/
+/*!	\brief Drop current record from queue
+**	\param [in,out] q - pointer of queue to handle
+**	\return drop status
+**	\retval true if successfully dropped from queue
+**	\retval false if queue is empty
+**/
 bool q_drop(Queue_t * q);
 
 
+/****************************************************************/
 #ifdef __cplusplus
 }
 #endif
 
-#endif /* __CQUEUE_H__ */
+#endif /* __CQUEUE_H */
+/****************************************************************/